fix msvc target (this has only been tested on clang for a long time).

+++ PKRenderer/Source/App/Engines/EngineProfiler.cpp
@@ -87,7 +87,7 @@ namespace PK::App
         {
             const auto s_offset = (i + sampleCountMax - sampleCountMin) * sampleWidth;
             const auto s_sample = m_timeHistory[(m_timeHistoryHead + i + 1ull) % sampleCountMin];
-            const auto s_normalized = (s_sample - minHistoryTime) / (maxHistoryTime - minHistoryTime);
+            const auto s_normalized = (float)((s_sample - minHistoryTime) / (maxHistoryTime - minHistoryTime));
             const auto s_height = (int)math::round(sampleHeight * s_normalized);
             const auto s_color = Math::HueToRGB32((1.0f - s_normalized) / 3.0f);
             gui->GUIDrawRect(color32(s_color.r, s_color.g, s_color.b, 127), rectSample + short4(s_offset, 0, 0, -s_height));
