+++ PKRenderer/Source/Core/Rendering/Font.h
@@ -52,8 +52,8 @@ namespace PK
         RHITexture* GetRHI();
         const RHITexture* GetRHI() const;
 
-        static size_t CalculateMaxRectCount(const char* text, const Font* font);
-        static size_t CalculateRects(const char* text, const Font* font, const short4& area_rect, const short4& clip_rect, const FontStyle& style, FontRect* out_rects, size_t max_rects);
+        static uint32_t CalculateMaxRectCount(const char* text, const Font* font);
+        static uint32_t CalculateRects(const char* text, const Font* font, const short4& area_rect, const short4& clip_rect, const FontStyle& style, FontRect* out_rects, uint32_t max_rects);
 
     private:
         RHITextureRef m_texture = nullptr;
