+++ PKRenderer/Source/Core/Utilities/FixedPool.h
@@ -104,7 +104,7 @@ namespace PK
         protected:
             T* Allocate(size_t count, int64_t index) final
             {
-                index = index != -1 ? index : m_mask.FindFirstZeroRange(count);
+                index = index != -1ll ? index : m_mask.FindFirstZeroRange((uint32_t)count);
                 Memory::Assert(index >= 0ll && index + count - 1ll < capacity, "Pool capacity exceeded!");
                 auto ptr = GetData() + index;
                 m_mask.FlipRange(index, index + count);
